requirements.txt
aiogram==2.24

python-dotenv~=1.0.0